MyCast/requirements.txt
Django==3.0.5
Flask==1.1.1
Flask-Caching==1.8.0
image==1.5.28
matplotlib==3.2.1
MetPy==0.12.0
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.0
Pint==0.11
scipy==1.4.1
siphon==0.8.0
